@@ -11,65 +11,57 @@ description: A new Flutter project.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 2.3.0+1
+version: 2.3.1+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.16.2 <3.0.0"
 
 dependencies:
+  agora_rtc_engine: ^5.2.0
+  another_flushbar: ^1.10.29
+  auto_size_text: ^3.0.0
+  cached_network_image: ^3.2.0
+  cupertino_icons: ^1.0.5
+  firebase_messaging: ^12.0.1
   flutter:
     sdk: flutter
-  agora_rtc_engine: ^3.4.6
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  # Firebase Cloud Messaging Package 
-  firebase_messaging: ^7.0.3
-  flutter_svg: ^0.19.0
-  flutter_swiper : ^1.1.6
-  url_launcher: ^5.7.10
-  shared_preferences: '>=0.5.12+2 <2.0.0'
-  http:
-  intl:
-  qr_flutter:
-  share_extend:
-  path_provider:
-  permission_handler: ^5.1.0+2
-  image:
   flutter_localizations:
     sdk: flutter
-  cached_network_image: ^2.0.0
-  # image_downloader: ^0.19.0
+  flutter_speed_dial: ^6.0.0
+  flutter_staggered_animations: ^1.0.0
+  flutter_styled_toast: ^2.0.0
+  flutter_svg: ^1.0.3
+  flutter_swiper_plus: ^2.0.4
+  http: null
+  image: null
   image_gallery_saver: ^1.2.2
-  table_calendar: ^2.3.3
-  pie_chart:
-  flushbar: ^1.10.4
-  # Este plugin solo trabaja con permisos para android
-  permissions_plugin: ^1.1.4
-  whatsapp_unilink: ^1.0.0+1
-  pinch_zoom_image_last: ^0.3.1
-  flutter_styled_toast: ^1.3.1
-  flutter_speed_dial: ^1.1.2
-  image_picker: ^0.6.7+22
-  auto_size_text: ^2.1.0
-  zoom_widget: ^0.1.3
-  dynamic_list_view: ^0.2.1
-  flutter_staggered_animations: ^0.1.3
-  
+  image_picker: ^0.8.5+3
+  intl: null
+  onesignal_flutter: ^3.3.2
+  pagination_view: ^2.0.0
+  path_provider: null
+  permission_handler: ^10.0.0
+  pie_chart: null
+  pinch_zoom_image_last: ^0.4.1
+  qr_flutter: null
+  share_extend: null
+  shared_preferences: ^2.0.15
+  table_calendar: ^3.0.6
+  url_launcher: ^6.1.5
+  whatsapp_unilink: ^2.0.2
+  zoom_widget: ^0.2.1
+
 dependency_overrides:
-    cached_network_image: ^2.3.3
-    
+  cached_network_image: ^2.3.3
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
-
 # The following section is specific to Flutter.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
