@@ -2,16 +2,6 @@ import 'dart:convert';
 
 import 'package:shared_preferences/shared_preferences.dart';
 
-/*
-  Recordar instalar el paquete de:
-    shared_preferences:
-  Inicializar en el main
-    final prefs = new PreferenciasUsuario();
-    await prefs.initPrefs();
-    
-    Recordar que el main() debe de ser async {...
-*/
-
 class PreferenciasUsuario {
   static final PreferenciasUsuario _instancia =
       new PreferenciasUsuario._internal();
@@ -22,7 +12,7 @@ class PreferenciasUsuario {
 
   PreferenciasUsuario._internal();
 
-  SharedPreferences _prefs;
+  late SharedPreferences _prefs;
   Codec<String, String> stringToBase64 = utf8.fuse(base64);
 
   initPrefs() async {
@@ -30,7 +20,7 @@ class PreferenciasUsuario {
   }
 
   // GET y SET del token
-  get token {
+  String get token {
     return _prefs.getString('token') ?? '';
   }
 
@@ -39,7 +29,7 @@ class PreferenciasUsuario {
   }
 
   // GET y SET del player ID
-  get playerID {
+  String get playerID {
     return _prefs.getString('playerID') ?? '';
   }
 
@@ -48,7 +38,7 @@ class PreferenciasUsuario {
   }
 
   // GET y SET del idUsuario
-  get usuarioLogged {
+  String get usuarioLogged {
     return _prefs.getString('usuario') ?? '';
   }
 
@@ -56,7 +46,7 @@ class PreferenciasUsuario {
     _prefs.setString('usuario', value);
   }
 
-  get themeMode {
+  String get themeMode {
     return _prefs.getString('tema') ?? 'Dark';
   }
 
