@@ -23,20 +23,19 @@ class _MainPageState extends State<MainPage> with WidgetsBindingObserver {
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((_) => Future.delayed(
+    WidgetsBinding.instance?.addPostFrameCallback((_) => Future.delayed(
         Duration(milliseconds: 2300), () => pushManager.mostrarUltimaVisita()));
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance?.addObserver(this);
     _validaSesion.sesion.listen((sesionValida) {
       if (sesionValida == 0) {
         if (navigatorKey.currentContext != null) {
           _prefs.borraPrefs();
-          //print('${_prefs.usuarioLogged}');
-          Navigator.of(navigatorKey.currentContext).pushNamedAndRemoveUntil(
+          Navigator.of(navigatorKey.currentContext!).pushNamedAndRemoveUntil(
               'login', (Route<dynamic> route) => false);
         }
       }
       if (sesionValida == 2) {
-        creaDialogBloqueo(navigatorKey.currentContext, 'Cuenta Suspendida',
+        creaDialogBloqueo(navigatorKey.currentContext!, 'Cuenta Suspendida',
             'Tu cuenta ha sido suspendida. Para reactivarla, comunícate con tu administración.');
       }
     });
@@ -46,14 +45,14 @@ class _MainPageState extends State<MainPage> with WidgetsBindingObserver {
             .creaSnackPersistent(
                 Icons.notifications_active, data['areas'], Colors.white,
                 dismissible: true)
-            .show(navigatorKey.currentContext);
+            .show(navigatorKey.currentContext!);
       if (data.containsKey('aviso'))
-        Navigator.pushNamed(navigatorKey.currentContext, 'AvisoDetalle',
+        Navigator.pushNamed(navigatorKey.currentContext!, 'AvisoDetalle',
             arguments: data['aviso']);
       if (data.containsKey('visita')) {
         ///previene la llamada del setState cuando el widget ya ha sido destruido. (if (!mounted) return;)
         if (!mounted) return;
-        await Navigator.pushNamed(navigatorKey.currentContext, 'VisitaNotif',
+        await Navigator.pushNamed(navigatorKey.currentContext!, 'VisitaNotif',
             arguments: data['visita']);
         setState(() {});
       }
@@ -95,6 +94,6 @@ class _MainPageState extends State<MainPage> with WidgetsBindingObserver {
   @override
   void dispose() {
     super.dispose();
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance?.removeObserver(this);
   }
 }
