@@ -1,6 +1,7 @@
+import 'dart:developer';
+
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:dostop_v2/src/widgets/elevated_container.dart';
-import 'package:dynamic_list_view/dynamic_list.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
@@ -9,6 +10,7 @@ import 'package:dostop_v2/src/providers/mis_accesos_provider.dart';
 import 'package:dostop_v2/src/utils/preferencias_usuario.dart';
 import 'package:dostop_v2/src/utils/dialogs.dart';
 import 'package:dostop_v2/src/utils/utils.dart' as utils;
+import 'package:pagination_view/pagination_view.dart';
 
 import 'package:pinch_zoom_image_last/pinch_zoom_image_last.dart';
 
@@ -23,14 +25,15 @@ class _MisAccesosPageState extends State<MisAccesosPage> {
   final _prefs = PreferenciasUsuario();
   final _key = GlobalKey();
   bool _notificarAccesos = false, _obteniendoConfig = true;
-  Future<List<AccesoModel>> _accesosProviderFuture;
-  int _pag = 1;
+  late int page;
+  late PaginationViewType paginationViewType;
+  late GlobalKey<PaginationViewState> key;
 
   @override
   void initState() {
     super.initState();
-    _accesosProviderFuture =
-        _accesosProvider.obtenerAccesos(_prefs.usuarioLogged, _pag);
+    paginationViewType = PaginationViewType.listView;
+    key = GlobalKey<PaginationViewState>();
     _configUsuarioProvider
         .obtenerEstadoConfig(_prefs.usuarioLogged, 3)
         .then((resultado) {
@@ -93,10 +96,6 @@ class _MisAccesosPageState extends State<MisAccesosPage> {
     Navigator.of(context).pop('dialog');
     setState(() {
       _notificarAccesos = resultado['OK'] == 1 ? valor : _notificarAccesos;
-      //   Scaffold.of(context).showSnackBar(utils.creaSnackBarIcon(
-      //       Icon(resultado['OK'] == 1 ? Icons.notifications_active : Icons.error),
-      //       resultado['message'],
-      //       5));
     });
   }
 
@@ -104,30 +103,29 @@ class _MisAccesosPageState extends State<MisAccesosPage> {
     return Container();
   }
 
+  Future<List<AccesoModel>> _dataRequester(int offset) async {
+    page = (offset / 10).ceil() + 1;
+    List<AccesoModel> list =
+        await _accesosProvider.obtenerAccesos(_prefs.usuarioLogged, page);
+    return list;
+  }
+
   Widget _cargaListadoAccesos() {
-    return DynamicList.build(
+    return PaginationView(
         key: _key,
-        dataRequester: _dataRequester,
-        initRequester: _initRequester,
-        itemBuilder: (List dataList, BuildContext context, int index) {
-          if (dataList.length > 0)
-            return Padding(
-                padding: EdgeInsets.only(bottom: 20),
-                child: _crearItem(context, dataList[index], index));
-          else
-            return Center(
-              child: Text('No se encontraron accesos'),
-            );
-        });
-  }
-
-  Future<List> _dataRequester() async {
-    _pag++;
-    return await _accesosProvider.obtenerAccesos(_prefs.usuarioLogged, _pag);
-  }
-
-  Future<List> _initRequester() async {
-    return Future.value(_accesosProviderFuture);
+        itemBuilder: (BuildContext context, AccesoModel acceso, int index) {
+          return Padding(
+              padding: EdgeInsets.only(bottom: 20),
+              child: _crearItem(context, acceso, index));
+        },
+        pullToRefresh: true,
+        pageFetch: _dataRequester,
+        onEmpty: const Center(
+          child: Text('No se encontraron accesos'),
+        ),
+        onError: (dynamic error) => const Center(
+              child: Text('Some error occured'),
+            ));
   }
 
   Widget _crearItem(BuildContext context, AccesoModel acceso, int index) {
