@@ -25,19 +25,20 @@ class LoginValidator {
   }
 
   dispose() {
-    _sesionStreamController?.close();
+    _sesionStreamController.close();
   }
 
-  Future<Null> verificaSesion() async {
+ Future<Null> verificaSesion() async {
     Map<String, dynamic> mapResp = Map<String, dynamic>();
     final authData = {
       'id': _prefs.usuarioLogged,
       'token': _prefs.token,
+      'playerID':_prefs.playerID,
       'fecha':
           DateFormat('yyyy-MM-dd HH:mm:ss').format(DateTime.now()).toString()
     };
     try {
-      final resp = await http.post('${constantes.urlApp}/validar_sesion2.php', body: authData);
+      final resp = await http.post(Uri.parse('${constantes.urlApp}/validar_sesion_os.php'), body: authData);
       mapResp = json.decode(resp.body);
       //decodeResp[0].forEach((String k, dynamic v) => mapResp[k] = v);
       if (resp.statusCode != 404 || resp.statusCode != 500) {
